+++ packages/@tailwindcss-standalone/src/types.d.ts
@@ -8,3 +8,4 @@ declare var __tw_resolve: undefined | ((id: string, base?: string) => string | f
 declare var __tw_readFile:
   | undefined
   | ((path: string, encoding: BufferEncoding) => Promise<string | undefined>)
+declare var __tw_load: undefined | ((path: string) => Promise<object | undefined>)